+++ CHANGELOG.md
@@ -13,6 +13,11 @@ CHANGELOG
   git grep --line-number --color=always -- '' |
       fzf --ansi --delimiter : --freeze-left 1 --keep-right
   ```
+- Also added `--freeze-right=N` option to keep the rightmost N columns visible.
+  ```sh
+  fd | fzf --freeze-right 1 --delimiter /
+  fd | fzf --freeze-left 1 --freeze-right 1 --delimiter /
+  ```
 
 0.66.1
 ------